requirements.txt
django
python-dotenv
hiredis
redis
requests
pydantic
django-components